SCRUM-70 feat: Refactor authentication hooks and enhance OAuth sign-in functionality
