+++ src/shared/composites/OAuthIcons/OAuthIcons.module.scss
@@ -16,3 +16,16 @@
     padding: 0;
   }
 }
+
+.btn {
+  padding: 0;
+  min-width: auto;
+  height: auto;
+  color: var(--color-light-100);
+
+  &:hover,
+  &:focus,
+  &:focus-visible {
+    color: var(--color-light-100);
+  }
+}
